backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
groq==0.4.2
sentence-transformers==2.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
httpx==0.26.0
pydantic==2.5.3
hf-xet==1.2.0
shellingham==1.5.4
typer-slim==0.24.0
rich==14.3.2
annotated-doc==0.0.4
markdown-it-py==4.0.0
pygments==2.19.2
mdurl==0.1.2
PyPDF2==3.0.1
